test: split cli suites by command
