+++ test/cli/params/test_registry.py
