+++ test/cli/rendering/test_log_view.py
