+++ test/cli/rendering/test_pretty.py
