+++ test/cli/rendering/test_progress.py
