+++ test/cli/workspace/test_workspace_cli.py
@@ -935,7 +935,9 @@ def test_run_flow_rerun_clears_states_and_stops_on_failure(monkeypatch, tmp_path
         engine_flow.prepared_for_rerun = True
         engine_flow.call_order.append(("prepare_rerun", workspace.directory))
 
-    monkeypatch.setattr("chipcompiler.data.prepare_workspace_for_rerun", prepare_workspace_for_rerun)
+    monkeypatch.setattr(
+        "chipcompiler.data.prepare_workspace_for_rerun", prepare_workspace_for_rerun
+    )
 
     rc = cli_main.run(["workspace", "run-flow", "--directory", str(ws), "--rerun", "--json"])
 
@@ -977,7 +979,9 @@ def test_run_flow_rerun_stops_when_prepare_fails(monkeypatch, tmp_path, capsys):
     def prepare_workspace_for_rerun(workspace, engine_flow):
         raise RuntimeError("cleanup failed")
 
-    monkeypatch.setattr("chipcompiler.data.prepare_workspace_for_rerun", prepare_workspace_for_rerun)
+    monkeypatch.setattr(
+        "chipcompiler.data.prepare_workspace_for_rerun", prepare_workspace_for_rerun
+    )
 
     rc = cli_main.run(["workspace", "run-flow", "--directory", str(ws), "--rerun", "--json"])
